(svn r21731) -Fix (r21728): don't forget to update the saveload version, or to check for MSVC warnings

+++ src/company_cmd.cpp
@@ -260,8 +260,8 @@ void UpdateLandscapingLimits()
 {
 	Company *c;
 	FOR_ALL_COMPANIES(c) {
-		c->terraform_limit = min(c->terraform_limit + _settings_game.construction.terraform_per_64k_frames, _settings_game.construction.terraform_frame_burst << 16);
-		c->clear_limit     = min(c->clear_limit     + _settings_game.construction.clear_per_64k_frames,     _settings_game.construction.clear_frame_burst << 16);
+		c->terraform_limit = min(c->terraform_limit + _settings_game.construction.terraform_per_64k_frames, (uint32)_settings_game.construction.terraform_frame_burst << 16);
+		c->clear_limit     = min(c->clear_limit     + _settings_game.construction.clear_per_64k_frames,     (uint32)_settings_game.construction.clear_frame_burst << 16);
 	}
 }
